# a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'askFEUP API'
  description: 'Web Resources Specification (A7) for askFEUP'
servers:
- url: http://lbaw.fe.up.pt
  description: Production server
...
paths:
...
  /login:
    get:
      operationId: R101
      summary: 'R101: Login Form'
      description: 'Provide login form. Access: GST'
      tags:
        - 'M01: Authentication'
      responses:
        '200':
          description: 'Ok. Show Log-in UI'

    post:
      operationId: R102
      summary: 'R102: Login Action'
      description: 'Processes the login form submission. Access: GST'
      tags:
        - 'M01: Authentication'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:         
                  type: string
                  format: email
                password:   
                  type: string
                  format: password
              required:
                  - email
                  - password
      responses:
        '302':
          description: 'Redirect after processing the login credentials.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful authentication. Redirect to home page.'
                  value: '/home'
                302Error:
                  description: 'Unsuccessful authentication. Redirect to login form.'
                  value: '/login'
...
  /questions/{id}:
    get:
      operationId: R303
      summary: 'R303: View a specific question' 
      description: 'Show a question, Access: USR'
      tags:
        - 'M03: Posts'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the question to retrieve.
      responses:
        '200':
          description: 'Success! Show the question.'
        '404':
          description: 'Question not found.'
        '403':
          description: "You don't have access to the question."
...